+++ Headers/countingSort.h
@@ -1,6 +1,6 @@
 #include<iostream>
 #include <vector>
 
-void countingSort(std::vector<int>& v){
+void countingSort(std::vector<double>& v){
 
 }
